+++ lib/type_safe_wait_on_futures.dart
@@ -1,43 +1,44 @@
 library type_safe_wait_on_futures;
 
-import 'package:rxdart/rxdart.dart';
-
-class FutureX {
+class Futures {
   static Future<(A, B)> wait2<A, B>(
     Future<A> a,
     Future<B> b,
-  ) =>
-      Rx.combineLatest2(
-        Stream.fromFuture(a),
-        Stream.fromFuture(b),
-        (a, b) => (a, b),
-      ).first;
+  ) async {
+    final res = await Future.wait([a, b]);
+    return (
+      res[0] as A,
+      res[1] as B,
+    );
+  }
 
   static Future<(A, B, C)> wait3<A, B, C>(
     Future<A> a,
     Future<B> b,
     Future<C> c,
-  ) =>
-      Rx.combineLatest3(
-        Stream.fromFuture(a),
-        Stream.fromFuture(b),
-        Stream.fromFuture(c),
-        (a, b, c) => (a, b, c),
-      ).first;
+  ) async {
+    final res = await Future.wait([a, b, c]);
+    return (
+      res[0] as A,
+      res[1] as B,
+      res[2] as C,
+    );
+  }
 
   static Future<(A, B, C, D)> wait4<A, B, C, D>(
     Future<A> a,
     Future<B> b,
     Future<C> c,
     Future<D> d,
-  ) =>
-      Rx.combineLatest4(
-        Stream.fromFuture(a),
-        Stream.fromFuture(b),
-        Stream.fromFuture(c),
-        Stream.fromFuture(d),
-        (a, b, c, d) => (a, b, c, d),
-      ).first;
+  ) async {
+    final res = await Future.wait([a, b, c, d]);
+    return (
+      res[0] as A,
+      res[1] as B,
+      res[2] as C,
+      res[3] as D,
+    );
+  }
 
   static Future<(A, B, C, D, E)> wait5<A, B, C, D, E>(
     Future<A> a,
@@ -45,15 +46,16 @@ class FutureX {
     Future<C> c,
     Future<D> d,
     Future<E> e,
-  ) =>
-      Rx.combineLatest5(
-        Stream.fromFuture(a),
-        Stream.fromFuture(b),
-        Stream.fromFuture(c),
-        Stream.fromFuture(d),
-        Stream.fromFuture(e),
-        (a, b, c, d, e) => (a, b, c, d, e),
-      ).first;
+  ) async {
+    final res = await Future.wait([a, b, c, d, e]);
+    return (
+      res[0] as A,
+      res[1] as B,
+      res[2] as C,
+      res[3] as D,
+      res[4] as E,
+    );
+  }
 
   static Future<(A, B, C, D, E, F)> wait6<A, B, C, D, E, F>(
     Future<A> a,
@@ -62,16 +64,17 @@ class FutureX {
     Future<D> d,
     Future<E> e,
     Future<F> f,
-  ) =>
-      Rx.combineLatest6(
-        Stream.fromFuture(a),
-        Stream.fromFuture(b),
-        Stream.fromFuture(c),
-        Stream.fromFuture(d),
-        Stream.fromFuture(e),
-        Stream.fromFuture(f),
-        (a, b, c, d, e, f) => (a, b, c, d, e, f),
-      ).first;
+  ) async {
+    final res = await Future.wait([a, b, c, d, e, f]);
+    return (
+      res[0] as A,
+      res[1] as B,
+      res[2] as C,
+      res[3] as D,
+      res[4] as E,
+      res[5] as F,
+    );
+  }
 
   static Future<(A, B, C, D, E, F, G)> wait7<A, B, C, D, E, F, G>(
     Future<A> a,
@@ -81,17 +84,18 @@ class FutureX {
     Future<E> e,
     Future<F> f,
     Future<G> g,
-  ) =>
-      Rx.combineLatest7(
-        Stream.fromFuture(a),
-        Stream.fromFuture(b),
-        Stream.fromFuture(c),
-        Stream.fromFuture(d),
-        Stream.fromFuture(e),
-        Stream.fromFuture(f),
-        Stream.fromFuture(g),
-        (a, b, c, d, e, f, g) => (a, b, c, d, e, f, g),
-      ).first;
+  ) async {
+    final res = await Future.wait([a, b, c, d, e, f, g]);
+    return (
+      res[0] as A,
+      res[1] as B,
+      res[2] as C,
+      res[3] as D,
+      res[4] as E,
+      res[5] as F,
+      res[6] as G,
+    );
+  }
 
   static Future<(A, B, C, D, E, F, G, H)> wait8<A, B, C, D, E, F, G, H>(
     Future<A> a,
@@ -102,18 +106,19 @@ class FutureX {
     Future<F> f,
     Future<G> g,
     Future<H> h,
-  ) =>
-      Rx.combineLatest8(
-        Stream.fromFuture(a),
-        Stream.fromFuture(b),
-        Stream.fromFuture(c),
-        Stream.fromFuture(d),
-        Stream.fromFuture(e),
-        Stream.fromFuture(f),
-        Stream.fromFuture(g),
-        Stream.fromFuture(h),
-        (a, b, c, d, e, f, g, h) => (a, b, c, d, e, f, g, h),
-      ).first;
+  ) async {
+    final res = await Future.wait([a, b, c, d, e, f, g, h]);
+    return (
+      res[0] as A,
+      res[1] as B,
+      res[2] as C,
+      res[3] as D,
+      res[4] as E,
+      res[5] as F,
+      res[6] as G,
+      res[7] as H,
+    );
+  }
 
   static Future<(A, B, C, D, E, F, G, H, I)> wait9<A, B, C, D, E, F, G, H, I>(
     Future<A> a,
@@ -125,17 +130,18 @@ class FutureX {
     Future<G> g,
     Future<H> h,
     Future<I> i,
-  ) =>
-      Rx.combineLatest9(
-        Stream.fromFuture(a),
-        Stream.fromFuture(b),
-        Stream.fromFuture(c),
-        Stream.fromFuture(d),
-        Stream.fromFuture(e),
-        Stream.fromFuture(f),
-        Stream.fromFuture(g),
-        Stream.fromFuture(h),
-        Stream.fromFuture(i),
-        (a, b, c, d, e, f, g, h, i) => (a, b, c, d, e, f, g, h, i),
-      ).first;
+  ) async {
+    final res = await Future.wait([a, b, c, d, e, f, g, h, i]);
+    return (
+      res[0] as A,
+      res[1] as B,
+      res[2] as C,
+      res[3] as D,
+      res[4] as E,
+      res[5] as F,
+      res[6] as G,
+      res[7] as H,
+      res[8] as I,
+    );
+  }
 }