+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.0
+
+- Remove dependency on rxdart.
+- Rename FutureX to Futures.
+- Update description and readme.
+
 ## 0.1.0
 
 - Initial version.